+++ PatternPaint/libblinky/lightbuddyuploader.h
@@ -15,7 +15,7 @@ public:
 
     bool storePatterns(BlinkyController &controller, QList<PatternWriter> &patternWriters);
     bool updateFirmware(BlinkyController &controller);
-    bool restoreFirmware(int timeout);
+    bool restoreFirmware(qint64 timeout);
     QString getErrorString() const;
 
     QList<PatternWriter::Encoding> getSupportedEncodings() const;
